requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
lightgbm
nuscenes-devkit
